Enforce Clang, enable Weverything and fix all warnings

+++ libwt/screen.c
@@ -6,7 +6,7 @@
 
 #define LOG_TAG "libwt/screen"
 
-struct screen *screen_new(int width, int height, int buffer_lines)
+struct screen *screen_new(unsigned int width, unsigned int height, unsigned int buffer_lines)
 {
     struct screen *screen = malloc(sizeof(struct screen));
     if (screen == NULL)
@@ -21,8 +21,8 @@ struct screen *screen_new(int width, int height, int buffer_lines)
     screen->buffer_lines = buffer_lines;
 
     memset(&(screen->pen), 0, sizeof(struct screen_pen));
-    screen->pen.fg = -1;
-    screen->pen.bg = -1;
+    screen->pen.fg = (unsigned) -1;
+    screen->pen.bg = (unsigned) -1;
 
     screen->lines = malloc(buffer_lines * sizeof(struct screen_line));
     if (screen->lines == NULL)
@@ -33,7 +33,7 @@ struct screen *screen_new(int width, int height, int buffer_lines)
 
     memset(screen->lines, 0, buffer_lines * sizeof(struct screen_line));
 
-    for (int i = 0; i < height; i++)
+    for (unsigned int i = 0; i < height; i++)
     {
         int rc = screen_line_realloc(&(screen->lines[i]), width);
         if (rc < 0)
@@ -54,7 +54,7 @@ void free_screen(struct screen *screen)
     {
         if (screen->lines)
         {
-            for (int i = 0; i < screen->buffer_lines; i++)
+            for (unsigned int i = 0; i < screen->buffer_lines; i++)
                 free(screen->lines[i].cells);
 
             free(screen->lines);
@@ -64,7 +64,7 @@ void free_screen(struct screen *screen)
     }
 }
 
-int screen_line_realloc(struct screen_line *line, int length)
+int screen_line_realloc(struct screen_line *line, unsigned int length)
 {
     struct screen_cell *new_cells = realloc(line->cells, length * sizeof(struct screen_cell));
     if (new_cells == NULL)
